# deployment/api/cyclingstats-openapi.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "CyclingStats API",
    "version": "1.0.0.0"
  },
  "paths": {
...
    "/users/{userId}/riders/{riderId}": {
      "post": {
        "tags": [
          "Users"
        ],
        "operationId": "Users_AddFavoriteRider",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          },
          {
            "name": "riderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 2
          }
        ],
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFavorideRiderRequest"
              }
            }
          },
          "required": true,
          "x-position": 3
        },
        "responses": {
          "401": {
            "description": "Not authorized to favorite riders.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "Rider favorited."
          }
        }
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "operationId": "Users_RemoveFavoriteRider",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 1
          },
          {
            "name": "riderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 2
          }
        ],
        "responses": {
          "401": {
            "description": "Not authorized to remove favorite riders.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "Rider removed."
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ProblemDetails": {
        "type": "object",
        "additionalProperties": {
          "nullable": true
        },
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "CreateFavorideRiderRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "comment": {
            "type": "string"
          }
        }
      },
...
    }
  }
}